+++ common/db/buckets.go
@@ -13,4 +13,5 @@ const (
     BucketPRIndexByRG  = "pr_index_by_rg_num"
     BucketWebhookRetries = "webhook_retries"
     BucketConfigHistory  = "config_history"
+    BucketAPIKeys        = "api_keys"
 )